fix: fix tests.
