+++ src/components/ProgressIndicator/index.js
@@ -46,7 +46,6 @@ export default class ProgressIndicator extends Component {
             children,
             currentStepName,
             onClick,
-            id,
         } = this.props;
         const context = {
             currentStepName,
@@ -57,7 +56,6 @@ export default class ProgressIndicator extends Component {
 
         return (
             <div
-                id={id}
                 className={this.getContainerClassNames()}
                 style={style}>
                 <ol className="rainbow-progress-indicator_list">
@@ -81,8 +79,6 @@ ProgressIndicator.propTypes = {
     style: PropTypes.object,
     /** The action triggered when the element is clicked. */
     onClick: PropTypes.func,
-    /** The id of the outer element. */
-    id: PropTypes.string,
     /**
      * This prop that should not be visible in the documentation.
      * @ignore
@@ -96,5 +92,4 @@ ProgressIndicator.defaultProps = {
     style: undefined,
     onClick: () => {},
     children: null,
-    id: undefined,
 };
